FDG-11899 update tests
