--- src/components/charts/bar/bar-component/bar-component.module.scss
@@ -1,3 +1,0 @@
-.bar {
-  cursor: pointer;
-}
