+++ src/components/charts/bar/bar.spec.js
@@ -1,24 +1,8 @@
-import React from 'react';
-import { render, fireEvent, waitFor } from '@testing-library/react';
+import React, { act } from 'react';
+import { fireEvent, render, waitFor } from '@testing-library/react';
 import BarGraph from './bar';
 import { staggeredData } from '../helpers/helpersData';
-import helpers from './helpers/helpers';
-
-jest.mock('@nivo/bar', () => ({
-  Bar: jest.fn(({ barComponent: BarComponent, data, indexBy, keys }) => {
-    return (
-      <div data-testid="mock-nivo-bar">
-        {data.map((item, index) => (
-          <div key={index}>{BarComponent && <BarComponent index={index} value={item[keys[0]]} data={item} />}</div>
-        ))}
-      </div>
-    );
-  }),
-}));
-
-jest.mock('./bar-component/bar-component', () => ({ index, value, handleTempValueChange }) => {
-  return <button data-testid={`custom-bar-trigger-${index}`} onClick={() => handleTempValueChange(index, value)} />;
-});
+import userEvent from '@testing-library/user-event';
 
 describe('BarGraph component', () => {
   it('does not render anything if invalid params are detected', () => {
@@ -36,8 +20,6 @@ describe('BarGraph component - Custom bar graph', () => {
   const mouseEnterPropSpy = jest.fn();
   const mockSetTempValue = jest.fn();
   const mockSetTempDate = jest.fn();
-  const mouseEnterSpy = jest.spyOn(helpers, 'mouseEnterEvent');
-  const mouseLeaveSpy = jest.spyOn(helpers, 'mouseLeaveEvent');
 
   const barGraphProps = {
     graphData: staggeredData,
@@ -50,14 +32,12 @@ describe('BarGraph component - Custom bar graph', () => {
     dateField: 'year',
   };
 
+  beforeAll(() => {
+    jest.useFakeTimers();
+  });
+
   beforeEach(() => {
     jest.clearAllMocks();
-    // needed so mouseLeave executes callback
-    mouseLeaveSpy.mockImplementation((cardId, callback) => {
-      if (callback && typeof callback === 'function') {
-        callback();
-      }
-    });
   });
 
   it('triggers the mouseEnter and mouseLeave events', () => {
@@ -65,30 +45,23 @@ describe('BarGraph component - Custom bar graph', () => {
     const { getByTestId } = render(<BarGraph {...barGraphProps} />);
     const chartContainer = getByTestId('barGraph');
     fireEvent.mouseEnter(chartContainer);
-    expect(mouseEnterSpy).toHaveBeenCalledTimes(1);
     expect(mouseEnterPropSpy).toHaveBeenCalled();
     fireEvent.mouseLeave(chartContainer);
-    expect(mouseLeaveSpy).toHaveBeenCalledTimes(1);
+    jest.runAllTimers();
+    expect(mockSetTempValue).toHaveBeenCalledWith(null);
+    expect(mockSetTempDate).toHaveBeenCalledWith(null);
   });
 
-  it('calls setTempValue and setTempDate with the correct values when the bar is clicked', async () => {
-    const { getByTestId } = render(<BarGraph {...barGraphProps} />);
-    const firstBar = getByTestId('custom-bar-trigger-0');
-    fireEvent.click(firstBar);
+  it('tiggers events on keyboard interaction', async () => {
+    jest.clearAllMocks();
+
+    render(<BarGraph {...barGraphProps} />);
+    act(() => {
+      userEvent.tab();
+    });
     await waitFor(() => {
       expect(mockSetTempValue).toHaveBeenCalledWith('500');
       expect(mockSetTempDate).toHaveBeenCalledWith('2017');
-    });
-  });
-
-  it('resets temp states when mouse leaves the bar', async () => {
-    const { getByTestId } = render(<BarGraph {...barGraphProps} />);
-    const chartContainer = getByTestId('barGraph');
-    fireEvent.mouseLeave(chartContainer);
-    expect(mouseLeaveSpy).toHaveBeenCalledTimes(1);
-    await waitFor(() => {
-      expect(mockSetTempValue).toHaveBeenCalledWith(null);
-      expect(mockSetTempDate).toHaveBeenCalledWith(null);
     });
   });
 });
