+++ src/components/home-highlight-cards/home-highlight-card/home-highlight-card.tsx
@@ -35,7 +35,7 @@ import { IDatasetApi } from '../../../models/IDatasetApi';
 import { IDataset } from '../../../models/IDataset';
 import { DatasetFieldDataType } from '../../../models/fdg-types';
 import { ariaLabels, formatCardValue } from '../home-highlight-cards-helper/home-highlight-cards-helper';
-import BarGraph from '../../charts/recharts-bar/recharts-bar';
+import BarGraph from '../../charts/bar/bar';
 import Sparkler from './sparkler/sparkler';
 import { IconProp } from '@fortawesome/fontawesome-svg-core';
 import { ga4DataLayerPush } from '../../../helpers/google-analytics/google-analytics-helper';
@@ -348,7 +348,6 @@ const HomeHighlightCard: FunctionComponent<HighlightCardProps> = ({ cardId, data
                   setTempValue={setTempValue}
                   setTempDate={setTempDate}
                   dateField={api ? api.dateField : ''}
-                  useCustomBarComponent
                   mouseEnter={barChartMouseEnter}
                 />
                 <div className={xAxis}>
